Update internal for 22Q3 release
update internal to commit-id: e2b0fde631fce349e0e3ad42b2a4d40ce7634a97

Type: Code Improvement

+++ src/tim/vx/internal/src/libnnext/vsi_nn_vxkernel.c
@@ -478,7 +478,7 @@ vsi_status vsi_nn_ClientNodePassParameters
     )
 {
     vsi_status status;
-    uint8_t i;
+    uint32_t i;
 
     status = VSI_FAILURE;
     for( i = 0; i < num; i++ )
